added the selection tool
to use it in the backend, check the class epie.gui.selection

if a selection exists, it's sent to the backend when a request is performed. To retrieve it in the back-end retreive the post variable "selection", it's an array('x', 'y', 'width', 'height') representing the selection (the values are scaled in the frontend in case the image is not at a 100% zoom so no need to do it in ezpublish).



git-svn-id: http://svn.projects.ez.no/epie@15 621ec555-2961-457b-befc-e0ae98695168